+++ src/main/java/com/T82/coupon/controller/CouponController.java
@@ -6,6 +6,7 @@ import com.T82.coupon.dto.request.CouponVerifyRequestDto;
 import com.T82.coupon.dto.response.CouponResponseDto;
 import com.T82.coupon.dto.response.CouponVerifyResponseDto;
 import com.T82.coupon.global.domain.dto.UserDto;
+import com.T82.coupon.service.CouponEventService;
 import com.T82.coupon.service.CouponService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -22,6 +23,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/coupons")
 public class CouponController {
     private final CouponService couponService;
+    private final CouponEventService couponEventService;
 
     /**
      * 쿠폰 생성
@@ -72,6 +74,6 @@ public class CouponController {
     @PostMapping("/events")
     @ResponseStatus(HttpStatus.OK)
     public void createCouponEvent(@RequestBody CouponEventRequestDto req) {
-        couponService.createCouponEvent(req);
+        couponEventService.createCouponEvent(req);
     }
 }
