+++ src/main/java/com/T82/coupon/service/CouponEventService.java
@@ -1,0 +1,7 @@
+package com.T82.coupon.service;
+
+import com.T82.coupon.dto.request.CouponEventRequestDto;
+
+public interface CouponEventService {
+    void createCouponEvent(CouponEventRequestDto req);
+}
