+++ src/main/java/com/T82/coupon/service/CouponService.java
@@ -22,5 +22,5 @@ public interface CouponService {
 
     CouponVerifyResponseDto verifyCoupons(CouponVerifyRequestDto req);
 
-    void createCouponEvent(CouponEventRequestDto req);
+
 }
